fixed the metadata issue with Profile page
